requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.13.1
requests==2.27.1
typing-extensions==3.10.0.2
responses==0.21.0
